Refine seeding files, load data into PostgreSQL database
