+++ .gitignore
@@ -1,1 +1,4 @@
+.DS_Store
+package-lock.json
+seedingFiles
 node_modules
